refactor(cart): CartItemContainer로 로직 분리 (#57)
CartItemContainer에서 CartItem 컴포넌트를 활용한 로직을 작성했습니다.
